Move registry to separate module
