+++ ipa_consistency_checker/plugins/counters.py
@@ -23,7 +23,8 @@ Checker plugins which works as counters - count number of entries in LDAP
 
 from __future__ import absolute_import
 
-from ipa_consistency_checker.plugins.absplugin import LDAPPlugin, Registry
+from ipa_consistency_checker.plugins.absplugin import LDAPPlugin
+from ipa_consistency_checker.registry import Registry
 
 register = Registry()
